[#51] Fix build error, split cascade func into file
